+++ familyhelper-webapi-stack/src/main/java/com/dwarfeng/familyhelper/webapi/stack/service/note/NoteBookResponseService.java
@@ -1,9 +1,6 @@
 package com.dwarfeng.familyhelper.webapi.stack.service.note;
 
-import com.dwarfeng.familyhelper.note.stack.bean.dto.NoteBookCreateInfo;
-import com.dwarfeng.familyhelper.note.stack.bean.dto.NoteBookPermissionRemoveInfo;
-import com.dwarfeng.familyhelper.note.stack.bean.dto.NoteBookPermissionUpsertInfo;
-import com.dwarfeng.familyhelper.note.stack.bean.dto.NoteBookUpdateInfo;
+import com.dwarfeng.familyhelper.note.stack.bean.dto.*;
 import com.dwarfeng.familyhelper.note.stack.bean.entity.NoteBook;
 import com.dwarfeng.familyhelper.webapi.stack.bean.disp.note.DispNoteBook;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
@@ -27,11 +24,29 @@ public interface NoteBookResponseService extends Service {
 
     PagedData<NoteBook> all(PagingInfo pagingInfo) throws ServiceException;
 
+    PagedData<NoteBook> userOwned(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException;
+
+    PagedData<NoteBook> userPermittedWithConditionDisplay(
+            StringIdKey accountKey, String pattern, boolean onlyFavored, PagingInfo pagingInfo
+    ) throws ServiceException;
+
+    PagedData<NoteBook> userOwnedWithConditionDisplay(
+            StringIdKey accountKey, String pattern, boolean onlyFavored, PagingInfo pagingInfo
+    ) throws ServiceException;
+
     DispNoteBook getDisp(LongIdKey key, StringIdKey inspectAccountKey) throws ServiceException;
 
-    PagedData<DispNoteBook> allPermittedDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException;
+    PagedData<DispNoteBook> allDisp(StringIdKey inspectAccountKey, PagingInfo pagingInfo) throws ServiceException;
 
-    PagedData<DispNoteBook> allOwnedDisp(StringIdKey accountKey, PagingInfo pagingInfo) throws ServiceException;
+    PagedData<DispNoteBook> userOwnedDisp(StringIdKey inspectAccountKey, PagingInfo pagingInfo) throws ServiceException;
+
+    PagedData<DispNoteBook> userPermittedWithConditionDisplayDisp(
+            StringIdKey inspectAccountKey, String pattern, boolean onlyFavored, PagingInfo pagingInfo
+    ) throws ServiceException;
+
+    PagedData<DispNoteBook> userOwnedWithConditionDisplayDisp(
+            StringIdKey inspectAccountKey, String pattern, boolean onlyFavored, PagingInfo pagingInfo
+    ) throws ServiceException;
 
     LongIdKey createNoteBook(StringIdKey userKey, NoteBookCreateInfo noteBookCreateInfo)
             throws ServiceException;
@@ -40,7 +55,11 @@ public interface NoteBookResponseService extends Service {
 
     void removeNoteBook(StringIdKey userKey, LongIdKey noteBookKey) throws ServiceException;
 
-    void upsertPermission(StringIdKey userKey, NoteBookPermissionUpsertInfo permissionUpsertInfo) throws ServiceException;
+    void upsertPermission(StringIdKey userKey, NoteBookPermissionUpsertInfo permissionUpsertInfo)
+            throws ServiceException;
 
-    void removePermission(StringIdKey userKey, NoteBookPermissionRemoveInfo permissionRemoveInfo) throws ServiceException;
+    void removePermission(StringIdKey userKey, NoteBookPermissionRemoveInfo permissionRemoveInfo)
+            throws ServiceException;
+
+    void changeFavored(StringIdKey operateUserKey, NoteBookFavoredChangeInfo info) throws ServiceException;
 }